refactor: collapse detect+generate into single transform functions
Remove centralConfig, securedClusterConfig, and all child structs
(storageConfig, monitoringConfig, exposureConfig). Detection logic
is inlined directly into TransformToCentral and TransformToSecuredCluster,
which now build the CR in a single pass without intermediate state.

Shared helpers (envVarValue, envVarIsTrue, detectCustomImages,
detectDeclarativeConfig, findVolume) moved to helpers.go.

Tests collapsed from 4 files per resource type to 1 each, testing
the Transform* functions directly.
